feat: Add user API
